% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthweight.R
\name{birthweightclass}
\alias{birthweightclass}
\title{Classifying Birthweights Using Provided Centile Values}
\usage{
birthweightclass(data, centiles, GAcols, CENTcolsB, CENTcolsG, CENTcolsU, dataBW, BWscale = NULL, dataGEND, pathname)
}
\arguments{
\item{data}{object of class "data.frame" with birthweight data.}

\item{centiles}{object of class "data.frame" with centiles data.}

\item{GAcols}{string. Character string with names of columns in \code{data} and \code{centiles} containing gestational age (GA) of newborn. The GA column name in \code{data} should be first in the string, followed by the GA column name in \code{centiles}.}

\item{CENTcolsB}{string. Character string with names of columns in \code{data} containing the nth lower centile values and nth upper centile values for male newborns. The column with nth lower centile values should be named first in the string, followed by the column with nth upper centile values.}

\item{CENTcolsG}{string. Character string with names of columns in \code{data} containing the nth lower centile values and nth upper centile values for female newborns. The column with nth lower centile values should be named first in the string, followed by the column with nth upper centile values.}

\item{CENTcolsU}{string. Character string with names of columns in \code{data} containing the nth lower centile values and nth upper centile values for unknown gender newborns. The column with nth lower centile values should be named first in the string, followed by the column with nth upper centile values.}

\item{dataBW}{character. Name of column in \code{data} with newborn birthweights.}

\item{BWscale}{character. Must be one of "grams" or "kilos", specifying the measurement scale used in \code{dataBW}. Default: NULL}

\item{dataGEND}{character. Name of column from \code{data} with newborn gender.}

\item{pathname}{character. File path to folder where the output data frame and table should be written.}
}
\value{
A list with a data frame containing all the columns in \code{data} in addition to a column with birthweight class, as well as a table containing the birthweight class frequencies.
}
\description{
Return a list with a data frame and table containing birthweight classes.
}
\details{
Appropriate conversions need to be done for birthweights not recorded in either grams or kilos. Newborn genders should be recorded as either "male" or "female" with the option of the first letter being capitalized.

Birthweights will be classified as either Small-for-Gestational Age (SGA), Large-for-Gestational Age (LGA) or Appropriate-for-Gestational Age (AGA) based on the gender-specific centile values provided in \code{centiles}.
}
\examples{
\dontrun{
if(interactive()){
# Obtain birthweight classes
bwclasses.list <- birthweightclass(data = data,
centiles = centiles, GAcols = c("GA column in data",
 "GA column in centiles"), CENTcolsB = c("nth lower centile male",
  "nth upper centile male"), CENTcolsG = c("nth lower centile female",
  "nth upper centile female"),
   CENTcolsU = c("nth lower centile unknown gender",
  "nth upper centile unknown gender"),
   dataBW = "birthweight", BWscale = "grams",
   dataGEND = "newborn gender", pathname = "output path name")

# Extract data set with new column indicating birthweight class
bwclasses.dat <- bwclasses.list[[1]]

# Extract table with birthweight class frequencies
bwclasses.tab <- bwclasses.list[[2]]
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
\author{
Bancy Ngatia
}
